Reinterpret data_ptr_t to uintptr_t

+++ src/execution/join_hashtable.cpp
@@ -435,11 +435,11 @@ idx_t JoinHashTable::PrepareKeys(DataChunk &keys, vector<TupleDataVectorFormat> 
 }
 
 static void StorePointer(const_data_ptr_t pointer, data_ptr_t target) {
-	Store<uint64_t>(reinterpret_cast<uint64_t>(pointer), target);
+	Store<uint64_t>(static_cast<uint64_t>(reinterpret_cast<uintptr_t>(pointer)), target);
 }
 
 static data_ptr_t LoadPointer(const_data_ptr_t source) {
-	return reinterpret_cast<data_ptr_t>(Load<uint64_t>(source));
+	return reinterpret_cast<data_ptr_t>(static_cast<uintptr_t>(Load<uint64_t>(source)));
 }
 
 //! If we consider to insert into an entry we expct to be empty, if it was filled in the meantime the insert will not
@@ -485,11 +485,11 @@ static inline data_ptr_t InsertRowToEntry(atomic<ht_entry_t> &entry, const data_
 		// if we are not in parallel mode, we can just do the operation without any checks
 		ht_entry_t current_entry = entry.load(std::memory_order_relaxed);
 		data_ptr_t current_row_pointer = current_entry.GetPointerOrNull();
-		Printer::PrintF("Storing pointer %llu", reinterpret_cast<uint64_t>(current_row_pointer));
+		Printer::PrintF("Storing pointer %llu", reinterpret_cast<uintptr_t>(current_row_pointer));
 		StorePointer(current_row_pointer, row_ptr_to_insert + pointer_offset);
-		Printer::PrintF("Trying to set pointer %llu, salt %llu", reinterpret_cast<uint64_t>(row_ptr_to_insert), salt);
+		Printer::PrintF("Trying to set pointer %llu, salt %llu", reinterpret_cast<uintptr_t>(row_ptr_to_insert), salt);
 		entry = ht_entry_t::GetDesiredEntry(row_ptr_to_insert, salt);
-		Printer::PrintF("Setting entry to pointer %llu, salt %llu", reinterpret_cast<uint64_t>(entry.load().GetPointer()), entry.load().GetSalt());
+		Printer::PrintF("Setting entry to pointer %llu, salt %llu", reinterpret_cast<uintptr_t>(entry.load().GetPointer()), entry.load().GetSalt());
 		return nullptr;
 	}
 }
